lang/eval: Move PlanningOracle to its own file
Continuing to split up some of the source files that have grown long over
ongoing experimentation, this is the start of splitting the config_plan.go
file into smaller parts.

For now only PlanningOracle gets its own file, but other types related to
planning might follow in later commits.
